fix: refactor-license-tracking
